Optimisation, fixes, deathrays target enemy players, no FF announce
Deathrays now treat pvp-enemy-players as if they were an NPC with roughly 1000 health.

Returned turrets (and now teslas) to the optimised isDifferentTeam check,with an added pvpSameTeam check which runs only on players (and w/ one of the NWInts cached)

Added a pvpsameteam check to friendly-fire tracking/announcing. You will no longer be marked a liability for killing enemy sweepers, or have announcer lines played.

metatable optimisation in GetAliveSweepers because that came up as expensive last time for some reason.

Fixed lua error when findRespawnBeacon called without a player input.