Invisible Respawn System
Director now allows individual respawn vectors not tied to a beacon to be defined. These will spawn the player in a droppod when activated.

These are team-associated (with 1/jcorp being treated as default). Beacons take priority.

I intend to use this to create respawns within a certain radius of an objective when completed. My reasoning for using proximity to objectives is to hopefully allow players to group-up when respawning, similar to how they would with beacons. This may have to be balanced with the risk of spawn-camping in future.

This system is flexible and may be used by other things in the future (e.g. distress beacon prefab).