Adjust turret resistances
Acid now deals 1.25x instead of 1.5x

Removed shock vulnerability to make them stronger against teslas & because I don't think any base game enemies use this damage type. (The default scaling is 0.75x instead of 1 for... some reason (going to chalk that up to early-mapsweepers-merekidor game design) so this is equivalent to havling the damage they deal to turrets).

Aim is to make dumping turrets into teslas a (more) viable way of clearing them in pvp, as well as a *slight* nerf to worker antlions.