Fix dropflare & C4 sending to all teams
Dropflare now properly prevents sending to all players. I forgot to add ply as an argument to the airdrop function, so it was getting a nil owner.

Timed explosives now only send to your team. I originally didn't do this under the belief that enemies should keep this as it's the main thing warning of timed explosives' threat. However giving away player's locations is a bigger issue currently. Eventually I might make this network to enemies only if they're within LoS, but for now this will do.