Burrower Guard Prototype
Added "Burrower Guard" antlion boss, which is substantially weaker than the other antlion guards, but smaller (with a smaller hull size to match), and able to burrow to a nearby node if their target is unreachable.

This guard is based on the episodic antlion guard and so also mimics their behaviours.

Added additional nodegraph focused functions, including: nearbyNodes (unused), nearestNodeInView, and nodeSplat.

node splat roughly allows getting nodes in a radius whilst being confined to whichever zone the position is in. This allows nodes on the same roof or in the same building as a target to be differentiated from ones that are just close-by.