Reworked datadownload
Added jcms.mapgen_MainZone() to replace all of those zonelist indexes with/clean-up the code.

Reworked data download. Removed the password search (phase 1), made pillars spread out (and variable count). The defense is now the entirety of the mission and is longer.

Added jcms.mapgen_AreaFlat() and replaced duplicate code in various places with it.
Moved flat check out of overlooking check (it's in turret and emplacements' stamps directly now)