Faction Prefab System & improved floorturret prefab
Set up faction prefabs system properly so that prefabs can be defined for a specific faction.

Gave combine the ability to spawn these, and marked turrets as a combine prefab (These are now guaranteed spawns and scale with difficulty)

Limited floor turrets to the main zone (*Could* be reverted at some point but it was too likely for them to spawn in inaccessible areas, dilluting their effect)

Disabled turrets for the first 30s of a mission, with pings to warn that they're about to re-activate.

Prevented floor turrets from spawning in too-thin areas

Fixed AreaFlat returning the reverse of what it was supposed to