Proper PvP end-game condition & blacklisted respawn beacons
Respawn beacons are now removed for PVP rounds by default (as the intended gameplay is for you to get them exclusively from objectives). They can be returned with a convar.

Fixed floor-turret lua error due to incorrect variable name

Fixed games ending when respawn vectors are present.

Made the mission-end condition pvp-compatible. This should function exactly the same in singleplayer/non-pvp. In PVP it should end the mission as soon as only one team is left.