Separate deathray non-track-during-formation behaviour into entity variable
Instead of directly checking if the game is in pvp mode, deathrays now have a stored value determining the behaviour they will use. This allows it to be configured as needed.

The pvp-mode check is now in the spawnmenu order instead, and mainframe has been modified to always use this behaviour (those deathrays not being player-spawned means there's no reason for them to immediately track, and this will help players avoid them more easily).