Orbital beam fixes
Fixed orbital-beams failing when deployed directly on the ceiling skybox. This is caused by the original starting position being shifted up by a normal vec, which causes it to go out of the world if used directly on the sky.
I have fixed this by adding a check for if the initial trace hits the sky (in orbital_fixed_outdoors), and just returning that if it does.

Fixed orbital beams not being usable when there is a valid sky pos if the first hit position wasn't the sky. The hitSky variable was misleadingly named (it was true if the first trace hit the sky, rather than if the sky was hit in general).

Optimise util_GetSky (doesn't create a normalised up vector, we already have a generic one to re-use).
Gave it 32k instead of 20k, as that's closer to the actual max height of a map.