Fix mistake in shop generation.
Shop rework broke shop scaling and resulted in it being possible for no shops whatsoever to spawn.

Specifically:

2x multiplier for <= 0.9x difficulty was turned into +1 for <=0.9x difficulty, which meant it was not map scaled and would have had very poor results on large maps.

Map scaling was then applied to the guaranteed 1x with no failsafes (meaning it gets floored to 0 if the map's too small), which then meant as soon as you hit round 2 you'd have the possibility of having no shops whatsoever.