Mineralguard spawning tweaks/balance
Mineralguard frequency now scales with mission duration in addition to ore progress. This should hopefully mitigate issues with maps/missions where a lot of ores spawn out-of-bounds.

Increased mineralguard health and made them getRowdy on spawning. This is to bring them a bit closer to normal guards in balance / behaviour.