Orecrate polish
Ore crate now has a "Hit with stunstick to release ore" infotarget when it has contents. This is necessary as this isn't otherwise communicated (and trice actually forgot how to do this during playtesting).

Made orecrate try to place its ores in free spaces around the crate. This is meant to stop them from spewing everywhere whilst hopefully preventing them spawning in walls / under displacements. It isn't hugely reliable and still spews everywhere but is at least an improvement.