Many small design tweaks
Vehicles now properly receive damage dealt to their owner. This means they can't be used to wait-out-radiation indefinitely without actually avoiding it (Would have been an issue for APCs as they are unarmed, and have less freedom of movement, meaning situations where both sides just wait for the other to leave their vehicle were possible).

Increased orecrate ejectioncooldown because trice was struggling with this, and because it shouldn't affect the rest of us too much.

Made datadownload and mining ops give respawns at random positions rather than near the terminal/refinery. I originally tied them to objective location because it improved the likelihood of spawning somewhere safe, and was difficult to camp because objectives are spread out and you'd be spawning at a different place each time. This doesn't apply to datadownload or mining ops however as they plant you in a single consistent place which is likely to be contested throughout the match. In a recent playtest this caused trice to die 3 or 4 times in a row, because there was a nest and radiation by his refinery. This is not an optimal experience (falls under player being helpless / having to sit and wait while they lose).

Increased PVP datadownload objective wait time somewhat. Almost every download during our playtests has been a success, so 4x was probably a bit excessive (even if it felt right when setting it up). Realistically this should probably scale with map characteristics in pvp (as that matters for players having to travel larger distances, but not for NPCs which can teleport closer) but I'm assuming I don't have time for that.