Diminishing returns on respawns
Respawn waves are worth less as the team gets more of them.

This is jointly meant to be a catch-up mechanic to ensure one team doesn't become completely unassailable (e.g. situations like we had on ZMod where one side had 12 respawns while the other were on their last man), and an economy tweak to bring missions with more objectives like violence flashpoints closer to the avereage, because they spammed out a lot of respawns (they're responsible for that 12 respawn surplus).