Various lua error fixes, buff upgradestation shield
Fix payload running think / erroring during mission generation

Fix recon erroring if jumping before touching the ground for the first time.

Increased upgrade station shield upgrade from 1.25x to 1.5x. This is consistently the last upgrade to be picked (probably at least partly because it's the most boring), and it isn't actually that large of a bonus. Should hopefully bring it closer to the other 2 in terms of value.