Bountymul player variable, don't track join/leave in aftergame
Added the ability to set a bounty multiplier on the player. This is mostly a toy for modders (and maybe slightly for me. I am cooking something schizophrenic).

Suppress join/leave messages in aftergame lobby screen. I really didn't want to do this but I've caved after seeing full lobbies spontaneously combust one too many times (and being told "oh nothing was wrong I just didn't want to play on an empty server" and/or "Had to go to bed" by every person I've asked who was present for one of those (and being present for one myself actually)). This makes it less obvious when someone else has hopped off after a match (so that others aren't prompted to leave too, causing a chain reaction). This is a little underhanded, but hopefully the increased activity will be a net-positive People can still see who is/isn't there when returning to lobby anyhow, they just aren't being actively reminded of it.