Drastically improve turret & emplacement generation
Reduced emplacement inaccuracy (it's still way too much even after my previous buff)

Fix a mistake in one of the mapgen demonstrate functions

Got rid of CheckOverlooking completely, made a new function that uses navmesh visibility data to calculate the most open/visible direction.

Made a function that uses the above to get a position/angle, with the option of making that position opposite the side the angle is facing (for floor turrets).

Emplacements and turrets now use this instead of checkoverlooking. This means they now reliably face the right way (and can generate in more places).

Removed greater-than-average visibility check from emplacement.
Gave emplacement a weighting function that prefers high-up, high-visibility areas that are close-ish to walls (i.e. not in the centre of an open field, which vis on its own would tend towards).