PvP tracers cvar, adjust team chat logic, scattershot team-coloured tracers
Added jcms_laser tracers to players' bullets in PvP through a convar which is off by default. This will make it more obvious when you're being attacked and from where. This does change the dynamic a fair bit so it's not on by default just yet as I want to playtest it. In theory it should improve the gameplay though as not knowing where you're being attacked from generally isn't a positive experience (and snipers are pretty strong in PvP because of this).
I've named the cvar jcms_pvptracer instead of jcms_pvptracers so that the set value doesn't save / stick around when I change the default to 1

Adjusted team chat logic to hopefully make it so that NPC teamchat in PvP behaves as you'd expect it to (talking to your teammates exclusively rather than to the other NPC players). Ditto for voice chat.

Made jcms_laser turn yellow/red based on PvP team. Necessary for the tracers change to work, but this also makes scattershots consistent with their team like shelling and bolters are.