PvP antirad callin
A crate that grants 60 seconds of radiation immunity, available exclusively in PvP.

Intended to help players relocate through radiation in PvP's endgame. The cooldown and high cost should mean this doesn't help much with waiting out.

Often the only way to survive if you're in a bad spot with radiation is to plant a shield-charger, which isn't ideal as it forces you to stay in one place (Keeping you *inside* the radiation, and making it hard to attack other players). This offers an alternative that incentivises actually leaving the radiation more.

This should also help a bit in cases where players droppod into the radiation but can't afford a vehicle or survive long enough to escape otherwise.