Hopefully fix mining ops crash???
I have no idea how this is happening but it seems like ores might be instantly exploding *during the same tick that they're created* and triggering even more to be spawned. I have no idea if any of the stuff I've done will fix whatever's wrong because I still don't really understand why this is happening or how to trigger it. But hopefully all of these extra layers of safety will do *something*.